Add Bitbucket to the list of providers in the Add/Edit PAT window (#1380)
Add Bitbucket to the list of providers in the Add/Edit PAT window. If the PAT is a Bitbucket SAAS oauth 2 token, Bitbucket is selected by default in the provider's dropdown.
Exclude the Bitbucket item in the providers dropdown, when creating a new token, because Bitbucket SAAS does not provide PATs.