Modify the namespace provision request not to retry if an SSL error o… (#1399)
Add a new Axios wrapper that retries if any error except the SSL Handshake error occurs. This is needed to catch the SSL error while provisioning a PAT secret. The provision request removes the token secret, so if retry the request, the error will not be thrown as no secret will be provisioned.